Avoid passing unneeded integrations
